Move ARG_TYPE and base dispatcher utilities to expr_common
