+++ pyomo/core/expr/numeric_expr.py
@@ -16,7 +16,6 @@ import operator
 
 logger = logging.getLogger('pyomo.core')
 
-from pyomo.common import enums
 from pyomo.common.dependencies import attempt_import
 from pyomo.common.deprecation import deprecated, relocated_module_attribute
 from pyomo.common.errors import PyomoException, DeveloperError
@@ -36,6 +35,11 @@ from pyomo.core.expr.expr_common import (
     _lt,
     _le,
     _eq,
+    _unary_op_dispatcher_type_mapping,
+    _binary_op_dispatcher_type_mapping,
+    _invalid,
+    _recast_mutable,
+    NUMERIC_ARG_TYPE as ARG_TYPE,
 )
 
 # Note: pyggyback on expr.base's use of attempt_import(visitor)
@@ -90,6 +94,12 @@ relocated_module_attribute(
     'sizeof_expression',
     'pyomo.core.expr.visitor.sizeof_expression',
     version='6.6.2',
+    f_globals=globals(),
+)
+relocated_module_attribute(
+    'register_arg_type',
+    'pyomo.core.expr.expr_common',
+    version='6.8.3.dev0',
     f_globals=globals(),
 )
 
@@ -1637,21 +1647,6 @@ def _decompose_linear_terms(expr, multiplier=1):
 #
 # -------------------------------------------------------
 
-
-class ARG_TYPE(enums.IntEnum):
-    MUTABLE = -2
-    ASNUMERIC = -1
-    INVALID = 0
-    NATIVE = 1
-    NPV = 2
-    PARAM = 3
-    VAR = 4
-    MONOMIAL = 5
-    LINEAR = 6
-    SUM = 7
-    OTHER = 8
-
-
 _known_arg_types = {}
 
 
@@ -1713,91 +1708,6 @@ def _categorize_arg_type(arg):
 
 def _categorize_arg_types(*args):
     return tuple(_categorize_arg_type(arg) for arg in args)
-
-
-def _invalid(*args):
-    return NotImplemented
-
-
-def _recast_mutable(expr):
-    expr.make_immutable()
-    if expr._nargs > 1:
-        return expr
-    elif not expr._nargs:
-        return 0
-    else:
-        return expr._args_[0]
-
-
-def _unary_op_dispatcher_type_mapping(dispatcher, updates):
-    #
-    # Special case (wrapping) operators
-    #
-    def _asnumeric(a):
-        a = a.as_numeric()
-        return dispatcher[a.__class__](a)
-
-    def _mutable(a):
-        a = _recast_mutable(a)
-        return dispatcher[a.__class__](a)
-
-    mapping = {
-        ARG_TYPE.ASNUMERIC: _asnumeric,
-        ARG_TYPE.MUTABLE: _mutable,
-        ARG_TYPE.INVALID: _invalid,
-    }
-
-    mapping.update(updates)
-    return mapping
-
-
-def _binary_op_dispatcher_type_mapping(dispatcher, updates):
-    #
-    # Special case (wrapping) operators
-    #
-    def _any_asnumeric(a, b):
-        b = b.as_numeric()
-        return dispatcher[a.__class__, b.__class__](a, b)
-
-    def _asnumeric_any(a, b):
-        a = a.as_numeric()
-        return dispatcher[a.__class__, b.__class__](a, b)
-
-    def _asnumeric_asnumeric(a, b):
-        a = a.as_numeric()
-        b = b.as_numeric()
-        return dispatcher[a.__class__, b.__class__](a, b)
-
-    def _any_mutable(a, b):
-        b = _recast_mutable(b)
-        return dispatcher[a.__class__, b.__class__](a, b)
-
-    def _mutable_any(a, b):
-        a = _recast_mutable(a)
-        return dispatcher[a.__class__, b.__class__](a, b)
-
-    def _mutable_mutable(a, b):
-        if a is b:
-            a = b = _recast_mutable(a)
-        else:
-            a = _recast_mutable(a)
-            b = _recast_mutable(b)
-        return dispatcher[a.__class__, b.__class__](a, b)
-
-    mapping = {}
-    mapping.update({(i, ARG_TYPE.ASNUMERIC): _any_asnumeric for i in ARG_TYPE})
-    mapping.update({(ARG_TYPE.ASNUMERIC, i): _asnumeric_any for i in ARG_TYPE})
-    mapping[ARG_TYPE.ASNUMERIC, ARG_TYPE.ASNUMERIC] = _asnumeric_asnumeric
-
-    mapping.update({(i, ARG_TYPE.MUTABLE): _any_mutable for i in ARG_TYPE})
-    mapping.update({(ARG_TYPE.MUTABLE, i): _mutable_any for i in ARG_TYPE})
-    mapping[ARG_TYPE.MUTABLE, ARG_TYPE.MUTABLE] = _mutable_mutable
-
-    mapping.update({(i, ARG_TYPE.INVALID): _invalid for i in ARG_TYPE})
-    mapping.update({(ARG_TYPE.INVALID, i): _invalid for i in ARG_TYPE})
-
-    mapping.update(updates)
-    return mapping
 
 
 #
